requirements.txt
pytube2
rich
toml
py7zr